[tests] Pull out common submodules
New modules: fixture, process, program, target, and thin
