+++ tests/cache_check.rs
@@ -3,9 +3,12 @@ use anyhow::Result;
 mod common;
 
 use common::common_args::*;
+use common::fixture::*;
 use common::input_arg::*;
+use common::process::*;
+use common::program::*;
+use common::target::*;
 use common::test_dir::*;
-use common::*;
 
 //------------------------------------------
 
@@ -48,7 +51,7 @@ impl<'a> Program<'a> for CacheCheck {
 
 impl<'a> InputProgram<'a> for CacheCheck {
     fn mk_valid_input(td: &mut TestDir) -> Result<std::path::PathBuf> {
-        mk_valid_md(td)
+        common::thin::mk_valid_md(td) // FIXME: create cache metadata
     }
 
     fn file_not_found() -> &'a str {
