+++ tests/common/common_args.rs
@@ -1,5 +1,9 @@
-use crate::common::*;
+use anyhow::Result;
+
 use thinp::version::tools_version;
+
+use crate::common::process::*;
+use crate::common::program::*;
 
 //------------------------------------------
 // help
