+++ tests/common/output_option.rs
@@ -1,4 +1,10 @@
-use crate::common::*;
+use anyhow::Result;
+
+use thinp::file_utils;
+
+use crate::common::process::*;
+use crate::common::program::*;
+use crate::common::test_dir::*;
 
 //-----------------------------------------
 // test invalid arguments
