+++ tests/thin_delta.rs
@@ -3,8 +3,11 @@ use anyhow::Result;
 mod common;
 
 use common::common_args::*;
+use common::process::*;
+use common::program::*;
+use common::target::*;
 use common::test_dir::*;
-use common::*;
+use common::thin::*;
 
 //------------------------------------------
